+++ src/main/java/appeng/client/render/BaseBlockRender.java
@@ -725,9 +725,9 @@ public class BaseBlockRender
 			return 0;
 
 		if ( offset > 0 )
-			return (double) uv / 16.0;
+			return uv / 16.0;
 
-		return (16.0 - (double) uv) / 16.0;
+		return (16.0 - uv) / 16.0;
 	}
 
 	public void renderTile(AEBaseBlock block, AEBaseTile tile, Tessellator tess, double x, double y, double z, float f, RenderBlocks renderer)
