+++ src/main/java/appeng/client/render/RenderBlocksWorkaround.java
@@ -310,10 +310,10 @@ public class RenderBlocksWorkaround extends RenderBlocks
 			{
 				Tessellator tessellator = Tessellator.instance;
 
-				double d3 = (double) par8Icon.getInterpolatedU( this.renderMinZ * 16.0D );
-				double d4 = (double) par8Icon.getInterpolatedU( this.renderMaxZ * 16.0D );
-				double d5 = (double) par8Icon.getInterpolatedV( 16.0D - this.renderMaxY * 16.0D );
-				double d6 = (double) par8Icon.getInterpolatedV( 16.0D - this.renderMinY * 16.0D );
+				double d3 = par8Icon.getInterpolatedU( this.renderMinZ * 16.0D );
+				double d4 = par8Icon.getInterpolatedU( this.renderMaxZ * 16.0D );
+				double d5 = par8Icon.getInterpolatedV( 16.0D - this.renderMaxY * 16.0D );
+				double d6 = par8Icon.getInterpolatedV( 16.0D - this.renderMinY * 16.0D );
 
 				double d11 = par2 + this.renderMinX;
 				double d12 = par4 + this.renderMinY;
@@ -364,10 +364,10 @@ public class RenderBlocksWorkaround extends RenderBlocks
 			{
 				Tessellator tessellator = Tessellator.instance;
 
-				double d3 = (double) par8Icon.getInterpolatedU( 16.0D - this.renderMinZ * 16.0D );
-				double d4 = (double) par8Icon.getInterpolatedU( 16.0D - this.renderMaxZ * 16.0D );
-				double d5 = (double) par8Icon.getInterpolatedV( 16.0D - this.renderMaxY * 16.0D );
-				double d6 = (double) par8Icon.getInterpolatedV( 16.0D - this.renderMinY * 16.0D );
+				double d3 = par8Icon.getInterpolatedU( 16.0D - this.renderMinZ * 16.0D );
+				double d4 = par8Icon.getInterpolatedU( 16.0D - this.renderMaxZ * 16.0D );
+				double d5 = par8Icon.getInterpolatedV( 16.0D - this.renderMaxY * 16.0D );
+				double d6 = par8Icon.getInterpolatedV( 16.0D - this.renderMinY * 16.0D );
 
 				double d11 = par2 + this.renderMaxX;
 				double d12 = par4 + this.renderMinY;
@@ -446,10 +446,10 @@ public class RenderBlocksWorkaround extends RenderBlocks
 			{
 				Tessellator tessellator = Tessellator.instance;
 
-				double d3 = (double) par8Icon.getInterpolatedU( this.renderMinX * 16.0D );
-				double d4 = (double) par8Icon.getInterpolatedU( this.renderMaxX * 16.0D );
-				double d5 = (double) par8Icon.getInterpolatedV( this.renderMinZ * 16.0D );
-				double d6 = (double) par8Icon.getInterpolatedV( this.renderMaxZ * 16.0D );
+				double d3 = par8Icon.getInterpolatedU( this.renderMinX * 16.0D );
+				double d4 = par8Icon.getInterpolatedU( this.renderMaxX * 16.0D );
+				double d5 = par8Icon.getInterpolatedV( this.renderMinZ * 16.0D );
+				double d6 = par8Icon.getInterpolatedV( this.renderMaxZ * 16.0D );
 
 				double d11 = par2 + this.renderMinX;
 				double d12 = par2 + this.renderMaxX;
@@ -500,10 +500,10 @@ public class RenderBlocksWorkaround extends RenderBlocks
 			{
 				Tessellator tessellator = Tessellator.instance;
 
-				double d3 = (double) par8Icon.getInterpolatedU( this.renderMinX * 16.0D );
-				double d4 = (double) par8Icon.getInterpolatedU( this.renderMaxX * 16.0D );
-				double d5 = (double) par8Icon.getInterpolatedV( this.renderMinZ * 16.0D );
-				double d6 = (double) par8Icon.getInterpolatedV( this.renderMaxZ * 16.0D );
+				double d3 = par8Icon.getInterpolatedU( this.renderMinX * 16.0D );
+				double d4 = par8Icon.getInterpolatedU( this.renderMaxX * 16.0D );
+				double d5 = par8Icon.getInterpolatedV( this.renderMinZ * 16.0D );
+				double d6 = par8Icon.getInterpolatedV( this.renderMaxZ * 16.0D );
 
 				double d11 = par2 + this.renderMinX;
 				double d12 = par2 + this.renderMaxX;
@@ -554,10 +554,10 @@ public class RenderBlocksWorkaround extends RenderBlocks
 			{
 				Tessellator tessellator = Tessellator.instance;
 
-				double d3 = (double) par8Icon.getInterpolatedU( 16.0D - this.renderMinX * 16.0D );
-				double d4 = (double) par8Icon.getInterpolatedU( 16.0D - this.renderMaxX * 16.0D );
-				double d5 = (double) par8Icon.getInterpolatedV( 16.0D - this.renderMaxY * 16.0D );
-				double d6 = (double) par8Icon.getInterpolatedV( 16.0D - this.renderMinY * 16.0D );
+				double d3 = par8Icon.getInterpolatedU( 16.0D - this.renderMinX * 16.0D );
+				double d4 = par8Icon.getInterpolatedU( 16.0D - this.renderMaxX * 16.0D );
+				double d5 = par8Icon.getInterpolatedV( 16.0D - this.renderMaxY * 16.0D );
+				double d6 = par8Icon.getInterpolatedV( 16.0D - this.renderMinY * 16.0D );
 
 				double d11 = par2 + this.renderMinX;
 				double d12 = par2 + this.renderMaxX;
@@ -608,10 +608,10 @@ public class RenderBlocksWorkaround extends RenderBlocks
 			{
 				Tessellator tessellator = Tessellator.instance;
 
-				double d3 = (double) par8Icon.getInterpolatedU( this.renderMinX * 16.0D );
-				double d4 = (double) par8Icon.getInterpolatedU( this.renderMaxX * 16.0D );
-				double d5 = (double) par8Icon.getInterpolatedV( 16.0D - this.renderMaxY * 16.0D );
-				double d6 = (double) par8Icon.getInterpolatedV( 16.0D - this.renderMinY * 16.0D );
+				double d3 = par8Icon.getInterpolatedU( this.renderMinX * 16.0D );
+				double d4 = par8Icon.getInterpolatedU( this.renderMaxX * 16.0D );
+				double d5 = par8Icon.getInterpolatedV( 16.0D - this.renderMaxY * 16.0D );
+				double d6 = par8Icon.getInterpolatedV( 16.0D - this.renderMinY * 16.0D );
 
 				double d11 = par2 + this.renderMinX;
 				double d12 = par2 + this.renderMaxX;
