+++ src/main/java/appeng/container/AEBaseContainer.java
@@ -590,7 +590,7 @@ public abstract class AEBaseContainer extends Container
 	{
 		if ( syncData.containsKey( idx ) )
 		{
-			syncData.get( idx ).update( (long) value );
+			syncData.get( idx ).update( value );
 			return;
 		}
