+++ src/main/java/appeng/entity/RenderTinyTNTPrimed.java
@@ -30,9 +30,9 @@ public class RenderTinyTNTPrimed extends Render
 		GL11.glTranslatef( (float) x, (float) y - 0.25f, (float) z );
 		float f2;
 
-		if ( (float) tnt.fuse - life + 1.0F < 10.0F )
+		if ( tnt.fuse - life + 1.0F < 10.0F )
 		{
-			f2 = 1.0F - ((float) tnt.fuse - life + 1.0F) / 10.0F;
+			f2 = 1.0F - (tnt.fuse - life + 1.0F) / 10.0F;
 
 			if ( f2 < 0.0F )
 			{
@@ -51,7 +51,7 @@ public class RenderTinyTNTPrimed extends Render
 		}
 
 		GL11.glScalef( 0.5f, 0.5f, 0.5f );
-		f2 = (1.0F - ((float) tnt.fuse - life + 1.0F) / 100.0F) * 0.8F;
+		f2 = (1.0F - (tnt.fuse - life + 1.0F) / 100.0F) * 0.8F;
 		this.bindEntityTexture( tnt );
 		this.blockRenderer.renderBlockAsItem( Blocks.tnt, 0, tnt.getBrightness( life ) );
